create a home page and move anime list to anime page
