+++ client/src/App.tsx
@@ -1,7 +1,7 @@
 import './App.css';
 import { Route, Routes } from 'react-router-dom';
 import { Home_page } from './pages/home_page';
-import { Anime_page } from './pages/anime_page';
+import { AnimeView_page } from './pages/anime_view_page';
 import { Watch_page } from './pages/watch_page';
 import { Studio_page } from './pages/studio_page';
 import { Studios_page } from './pages/studios_page';
@@ -10,6 +10,7 @@ import { Genre_page } from './pages/genre_page';
 import { Manga_page } from './pages/manga_page';
 import { MangaView_page } from './pages/manga_view_page';
 import { MangaRead_page } from './pages/manga_read_page';
+import { Anime_page } from './pages/anime_page';
 
 function App() {
     return (
@@ -19,7 +20,8 @@ function App() {
             <Route path='/studios/:name' element={<Studio_page />} />
             <Route path='/genres' element={<Genres_page />} />
             <Route path='/genres/:name' element={<Genre_page />} />
-            <Route path='/anime/:title' element={<Anime_page />} />
+            <Route path='/anime' element={<Anime_page />} />
+            <Route path='/anime/:title' element={<AnimeView_page />} />
             <Route
                 path='/anime/:title/watch/:type/:number'
                 element={<Watch_page />}
