+++ client/src/pages/home_page.tsx
@@ -1,6 +1,9 @@
-import AnimeList from '../components/animeList';
 import '/public/css/pages/home.css';
 
 export function Home_page() {
-    return <AnimeList />;
+    return (
+        <>
+            hi
+        </>
+    )
 }
